requirements.txt
pandas~=1.4.3
plotly~=5.10.0
PyPDF2~=2.10.2
PyYAML~=6.0